"        tev2 d v i <| fun v i ->             match v, i with             | TyVV(l,_), TyLit (LitInt32 i as i') -> (fun _ -> l.[i]) |> f i l             | v & TyType (VVT ts), TyLit (LitInt32 i as i') -> (fun _ -> TyOp(VVIndex,[v;TyLit i'],ts.[i])) |> f i ts             | v, TyLit (LitInt32 i as i') -> d.on_type_er d.trace <| sprintf "Type of a evaluated expression in tuple index is not a tuple.\nGot: %A" v             | v, i -> d.on_type_er d.trace <| sprintf "Index into a tuple must be a natural number less than the size of the tuple.\nGot: %A" i
Ah, wow, so you can use the and patterns like this. And to think I had nothing more to learn about functional programming. What a fool I am.

This is actually super useful, far better than `as`. In combination with active patterns I do not need when at all."

I had an epiphany about and patterns.