I am done with the VVSliceFrom, EqType and TypeLitCreate cases.
With that only the two CaseTuple and CaseCons cases are left. As it 5pm almost, I think I will leave starting work on the for tomorrow.

It is really one case though since like VVSliceFrom derives from VVIndex with very minor changes, onceI have the CaseTuple, CaseCons will just drop out from it.

Just one more case and I will the full Spiral langauge at my fingertips. And the design for the case is already in my head, it is just matter of working it out in code.

But like the pattern match compiler it will be intricate and I do not want to rush into it. I want to think every little bit of the code through.

Nice and easy is the way to do it.