Yesterday, I talked that removing CPS from the language might be a good idea. Given how brilliantly the case function turned out to be, I am thinking of getting rid of CSE as well. I thought I would need it to propagate information, but I had no anticipated that the code would turn out like this once I started coding.
Really, I did not highlight it yesterday, but there are no words for how brilliantly the partially evaluated pattern matcher turned out. It is sheer genius that I can hardly believe I wrote it myself, especially considering how heavy the implementation of it in other languages are.
There is no doubt about it, Spiral is a lot closer to the True Language than any of the other languages. Optimization is understanding.

The reason why originally put in CPS is for the pattern matcher, but now that I am compiling the Pattern to Expr, the reason for that goes completely down the drain. I can return type errors using exception and removing CPS would speed up and simplify the pevaller significantly. CSE as well is a huge drain on the potential compilation speed of the pevaller. Not right now, but if I added dirty flags to TyEnv and TyVV to stop the destructure from pointlessly traversing the already destructured segments, having CSE would lead to incorrect code.

And it is a definitely a reason for slowdown right now.

I am not just making Spiral for myself, but for machines as well so I am going to have to be mindful of compilation times.

For today I am going to investigate the options and the languages mentioned by SirNastyPants in this post: https://www.reddit.com/r/ProgrammingLanguages/comments/6kn4ii/how_to_interface_a_language_with_the_net/djnx8yc/?st=j4m8yusr&sh=31fe7219

I'll stop with programming for the time being until I am able to draw out some kind of integration plan between Spiral and .NET.