Just now while I was thinking about whether to take out common subexpression elimination (CSE) as mentioned in the previous commit, I reallized that I in fact needed CSE for the pattern matcher as originally envisioned.
This happens to me sometimes - I'd forget the reasons for having a particular feature in the langauge and have to retrace the inference steps for it.

In this case though, I've realized that in fact as I'd written the thing, CSE would not rewrite variables so I've had to make some changes to the destructure function in order to make it flow correctly.

Right now, my agenda for the rest of the day is to read the papers on the Terra language and then take a look at the source for Mond. After that I'll try figuring out how dynamic langauges call .NET libraries.