Got rid of some useless `rec`s.
The Terra papers have been quite interesting so far. The approach they take is definitely not partial evaluation, and not as efficient as a language based on it, but regardless it is still very sensible approach to designing a language.

If Terra had a GPU backend and pattern matching I might seriously have considered taking it up.

But I think ultimately I can do a much better job with Spiral.