http://compilers.cs.uni-saarland.de/papers/gpce15.pdf Shallow Embedding of DSLs via Online Partial Evaluation
Specialization through Dynamic Staging
http://compilers.cs.uni-saarland.de/papers/gpce14.pdf

https://arxiv.org/abs/1612.01325
Building Code with Dynamic Staging

Here is a sample of the reading I did today so far. I was going to code today, but then decided to read Rompf Phd thesis.

https://infoscience.epfl.ch/record/180642/files/EPFL_TH5456.pdf
Lightweight Modular Staging and Embedded Compilers- Abstraction without Regret for High-Level High-Performance Programming

I am still at the beginning to it. I feel a lot lighter today than yesterday. The days started out similarly in that I woke up at around 7am which is pretty early for me, but whereas yesterday I had a tortured night, today I have a very good night's rest.

Somehow it feels like that bit by bit my confidence is coming back.

I am starting to internalize the belief that the language is really nearing completion and that I will in fact be able to use is to achieve my goals.

The more I read the staging literature, the more I am realizing that in fact, Spiral is unique.

I thought it might be dynamic staging, but that has nothing to do with it.

Thanks to the first paper though, I learned of a new term called first-class staging.

`Spiral - language with first-class staging and intensional polymorphism` is the perfect description for it.

First-class staging is perfect. It perfectly describes Spiral.

With it, I will be able to do it. Optimization is understanding and with such a powerful language, I will be able to combine DSLs with ease. With such a power I will be able to abstract away anything.

With such a power I will finally complete my programming education.

I am starting to look forward to it. I had enough of being a no-lifer.

Since Rompf's thesis is 191 pages, I'll take the rest of the day to read it. I'll start coding hopefully tomorrow. Tomorrow I will start by checking out DLR. I'll also check out Roslyn, though I expect to be generating F# code instead.