It occurred to me that I can compile the active patterns without introducing an intermediate let.
At any rate, I am done with Rompf's thesis.

The work on LMS is quite impressive. In my updates I praised the MLs in favor of Scala, but I guess in hindsight, Scala would have been the correct choice as it has good staging.

Right now I am too far gone to switch. I am nearly done with the language and it is time for the final push.

It is time to do .NET integration.