I've studied the Python source for SymPL for a while now. At first it seemed impossibly complicated, but now that I've studied it for a while I managed to internalize it a bit.
And that allowed me to see that 90% of the complexity in the SymPL implementation is really just boilerplate.

As I looked at the source for it, gradually the feeling that formed within me was cold rage and disdain. How dare they try to pass of this tripe as a serious compiler? Do they get paid by the word written?

As much as I am thrilled by interoprability of dynamic languages, there is no way I can use the half finished turd that is DLR.

If I was doing SymPL, it would have not been over 400 lines of code.

And while boilerplate might be boilerplate, getting it wrong will still lead to bugs and it is a weight that I need to carry in mind. It is unacceptable.

So where does that leave me? I guess I'll look at the other options /u/SirNastyPants mentioned.

Also going through the SymPL's source did remind me of something. The very first Cuda code generator that I tried, before I started work on Spiral, was in fact based on F# code quotations. That gave me some experience with manipulating .NET `Expression`s directly. If nothing else, I could always revisit that. Also Petzold's book gave me some experience with .NET's reflective capabilities as well.

...As useless as F# quotations are for making programs with them, they will be quite handy for exploring .NET internals.

Well, for now since it is the next thing on my list, let me check out that Roslyn tutorial.