I have no idea how to make the damn GetMethod and InvokeMember work on CreateInstance. I had enough for the day. I'll deal with this tomorrow.
To be honest, I am debating with myself whether I should do first class .NET integration. Maybe I should just use a special Op for that and do it through an interface instead.

Well, apart from knowing how to instantiate generics...no, I definitely do not have enough knowledge to start work on the code generator just yet. I need to study more.

I'll make .NET integration second class, the benefit of that is that it will allow me to branch out to other languages as needed.

But of what I do know, I know how to do generate F# programatically for my DSL. I'll use F#'s scripting compiler service to do it just as demonstrated in the compiler services documentation. Though before I decide on that, it might be better to generate a typechecked AST instead. That would allow me to use things like ValueTuples natively.

Tomorrow, the main thing on the agenda will be to get a grip on reflection and dynamic programming using it. No doubt, this will take me a while, but the important thing is to put in effort each and every day.