I did quite a bit of experimentation with .NET reflection just now. I figured out how to intantiate the type properly too.
With this, everything is set for me to make a .NET interface for my language. I won't allow creating .NET typees inside it, but I will allow using them.

Hmmmm...is there anything I am missing here?

Well, as is usual, after I have had a nice knowledge meal, it is best to rest a bit so it settles down in the mind.

So I will do that here.

The truth is, I feel like I still need to do more research on dynamic languages