I did `load_assembly` and `get_type_from_assembly`, but now that the focus turns to `instantiate_generic_params` I find that I need a model for going back and forth between Spiral and .NET types.
In theory, I could go a bit further despite it being 5:30pm, but not without a plan. Figuring out the intricacies of translating between the two languages is something I am going to have to dwell on.. Maybe it will turn out that I will decide to just build in the types I need. Probably not, but I'll let it float for a bit.

Let me call it a day here.

As a personal aside, for the past two months or so, I've been playing Bayonetta. I've put in an hour or two per day playing it, and now I find that I can beat it on Infinite Climax. I have 6 chapters left until I complete the run.

That is pretty amazing. After finishing the Hard Pure Platinum run, I thought the game would be boring, but Infinite Climax turns out is pretty exhilirating. At this point I am good enough at the game to actually enjoy it like this.

That is quite something.

Well, let me take my rest.