Before I start programming for the day, let me collect my thoughts. I've let them simmer enough. I've gone though the LMS tutorial, but it is just a reprint of Rompf's thesis and the LMS libraries seems to be abandonware at this point, which is a pity.
LMS does not matter, Spiral does.

https://www.youtube.com/watch?v=lC5UWG5N8oY&feature=youtu.be

I've also watched the above video on typesafe parallelism.

Here is what I am going to do, I will complete the basic .NET integration as planned.

Last night I had a bit of insight that made me doubt my current course - if I do .NET integration rather than building relevant functions and classes into the language, I am going to run into difficulty when it comes to doing the final step in Spiral: Dynamic -> Static transformation.

The way that transformation step would work is that the program itself would invoke the Spiral compiler and pass it the argument. The net effect of this is that since the program at runtime obviously knows the static value of every variable, the Spiral compiler will be able to do things impossible otherwise.

Such a technique would be called metacircular compilation or metacompilation.

If you think about it, that is not that much different from what current tracing JITs do except at a much smaller scale. And it is not that much different from the having the eval function in Javascript or Scheme per se.

But metacircular compilation is definitely intricate and not something that I can just undertake at this point.

I was not going to anyway, but it only occurred me yesterday that having .NET objects in the language would block this particular feature for the reason that I won't be able to automatically make static versions of .NET objects.

But on the other hand not having access to the .NET and having to build everything in would be a significant burden. Not to mention, I do not need the metacompilation right now, nor will I need it for the foreseeable.

Eventually I will though. In the future, in order to realize the true power of AI, I will need the language to have program synthesis capabilities. And program synthesis will in fact need metacompilation.

There are various difficulties with realizing metacompilation. In the first place, right now, I am compiling using strings.  That is how I am doing Cuda C++ and how I will target F#.

From the perspective of compilation speed, this is absolutely unacceptable. Cuda modules take 0.35s on average to compile even for very simple ones, and F# with all of its passes will not fare much better.

If machines are to use my language, I need to go lower. I need to directly target PTX for Cuda and .NET IL for the host side. Honestly, this is work I am not willing to put in right now.

I am yet to even make a single dollar despite working on this for so long; it is simply not worth considering until have AI agents with some measure of planning ability that could be then channeled into programming ability.

Since I've been away from working on the Spiral library for 9 months, it is worth reminding myself why the hell am I even working on this language - to give me generic Cuda map modules. Just that. To give me some abstraction capability in the face of C++ and to optimize away language crossing boundaries.

...

So, let me complete the first Spiral without metacompilation capability. Let me get some programming experience with the language, so that when the time comes to get rid of direct .NET interop I know which classes to bake in and which to leave out.

Actually, there won't be a need to get rid of .NET interop, but there will be a need to prevent the types from crossing metacompilation boundaries.

So I guess this is fine.

What I will do today is make that type converter and the rest of the cases.

After that, I will do the F# backend and complete the Spiral ML library. Then I will finally get into the game.

It is done. The language is basically done at this point. There is no need to agonize too much about interop. It is not like I have any other users I am beholden to, nor is it likely I have them for years to come.

It is time to finish the job and get the money.