I made all the functions necessary for .NET interop and plugged in the cases. Though instead of Op, I should have put them in into apply. Well, maybe. To start off, it is better that I be explicit.
At any rate, with this I have everything necessary for Spiral to interface with the .NET platform.

I guess I can start work on the code generator now.

Though writting this kind of boilerplatey interop code really comes hard to me. I do not feel like starting another thing right now considering how little there is left in the day.

Ah, yes, I also need to do the parser.

I've been thinking of moving back to the standard notation, but I am not exactly sure.

...Let me just stay like this. It is not that I mind Lisp syntax in pattern matching.

I guess I'll plug the rest of the parser back in and call it a day with that.