I finally plugged in the two ModuleWith cases. Previously I did the functions for them, but I decided to completely redesign them to take advantage of type level strings.
I really like my decision to remove named tuples and instead introduce type level strings.

I also renamed `fun`s to `met`s.

Is there anything I am missing? Well, I still need to put in unary operators. But that is just negation so I haven't bothered with it.

Also I would like to allow the users to be able to introduce their own operators at some point like with Haskell.

To be honest, this is such a pain in the ass, even now I am thinking whether I should just switch to the Lisp syntax and eliminate a whole host of syntax troubles.

Maybe I should deal with precedence the same way F# does?

The thing about programming is that it is not math despite it borowing from it. So yeah +,-,*,/ and the comparison operators are optimized for readability, but programming has to deal with everything else as well.

Well, on the bright side, machines will interface with code directly and won't have to deal with this nonsense.

...I think this is it with regards to parsing for the time being.

The next task is something that was long in coming - the F# backend. I am finally at this point.

As usual, let me rest for a bit before starting. I need to think about it first. I've been so focused on the parser that I did not spare much thought for the next step.