""" inl rec print_main buffer =     inl add = the ProgramNode >> buffer.Add     inl r = print_main buffer     function     | .state x -> .Statement x |> add     | .enter' f ->         .Indent |> add         f()         .Dedent |> add     | .enter f ->         r .enter' <| inl _ ->             match f() with             | "" -> ()             | s -> state s """
I feel the need to refactor the codegen now that I am doing it from scratch. The above is how the print_main would look like in Spiral.

The parser would not support this kind of thing currently. Let me make it so.

Though I'd want to do it as soon as possible, the codegen is in fact going to take me a few days to do correctly.

After that, I will get my first taste of real programming in Spiral. There is so much I could do with those Cuda kernels previously.

It should be a fun experiment.