While I was in the shower it occurred to me that I did the `inlr'` function completely wrong. What a fool I was today. Not only did I completely forget how to do folds properly, I also messed up the simple function. At any rate, the current version is a lot simpler and most importantly correct.
With this I really am done for the day.

https://github.com/AnthonyCalandra/modern-cpp-features

Right now this is what I am reading. Modern C++ is trying to turn itself into a staging language it seems. It have done it from the start. It won't manage to beat Spiral while dragging so much baggage behind it.