Ahhh, I fucked up.
I spent all this time during the morning designing the type_filter_unit and it did not even occur to me that the function would break the ordering structural guarantees that the code generator needs to function correctly.

I need to get rid of it as it is useless.

Instead....what I will do is go with the same idea I had in the Cuda backend, except do it properly this time.

What I failed to do correctly then is that I sort of tacked the unit checking onto all the functions that required it and did not bother to make sure the foundation of it was sound.

This time I am going to structure it correctly. There are a bunch of things I will need to change in the Cuda backend anyway, so it best that I test them out here first.