This is my last update for the day. I simply do not feel like it anymore. I've been at it on and off since 7:30am.
I'll finish the rest tomorrow and then get on with writing actual code in Spiral. I'll still have to put in union types and closures after that, which will take me a while. I am not even close to being done.

But I am getting closer.

In a sense, I am a different person from last year. I no longer have the drive to do this 12+ hours a day.

10h is enough for me.