Returning methods from methods works as I remember it. The recursion test surprised me, it seems now that static variables crash any kind of recursion. I'll add a Op to make vars dynamic without having to go through met.
Well, the next thing on the list is a feature that was long in coming. Union types and recursive union types. It is time to finally put them in for good. They are already in the pevaller and now it is time I figure out how to deal with them in the codegen.

That will be my task for today.

Well, I suppose before I starts, I could also test active patterns.