In the last commit, nearly 3 hours ago, I said that I was going to go on, but in truth I felt completely out of energy and decided to take a break. I literally could not push another thought.
What I did instead was surf the Internet for a bit and play Bayonetta. Amazingly enough, I managed to beat the Angel Slayer level! This was only my second try. I actually did not expect to win at all, I just wanted to waste some time and unwind, but somehow I got all the way to the 51 verse - the one against Bayonetta.

I had literally only 2HP, less than a tenth of a life bar left while she had 5 full life bars. I expected to die, but by taking advantage of Pulley's Butterfly and refilling HP to 2 every time it fell to 1 (the game has a mechanic that a character may only die on a hit from 1HP) using Star of Dinetta, I managed to whittle her down little by little.

When I noticed at some point that Bayonetta had only 1.5 life bars left, I started to sweat. I actually started feeling fear and anxiety that two hits in a row would be all it takes to end the run. But I stuck to my plan and just kept playing as best as I could.

In the end, to my great surprise I won! It was such a great battle.

The whole level took me nearly half an hour. If I had lost, I'd have to start it all the way from the beginning; unlike the regular chapters, there are no save points throughout.

And now that I've rested a bit and feeling better, I took another look at the TyOp Case which I was supposed to be working on.

```
        | TyOp(Case,v :: cases,_) ->
            sprintf "match %s with" (codegen v) |> state
            let rec loop = function
                | case :: body :: rest ->
                    sprintf "| %s ->" (codegen case) |> state
                    enter <| fun _ -> codegen body
                | [] -> ()
                | _ -> failwith "The cases should always be in pairs."
            loop cases
            ""
```
I wrote this out myself, but in the end I still came out surprised at how brief it came out. Now that I have a suply of motivation and am actually able to trace out the code paths, I find that it checks out.

I had forgotten just how similar binding and instantiation is in F#. In the Cuda backend it will be a lot more complicated, no doubt about it. But not particarly harder though.

3h ago it felt like it would take me an entire day to do it. In the end it was just 10m of work.

With this I've plugged pattern matching into the code generator. Wonderful.

This along with .NET interop were the reason for the new version of the language. Actually, now that I mention it, I still need to do .NET object instantiation, but that will be straightforward.

In terms of completeness, the language now has everything it needs to be a full language. From here on out, I will spend the next few week writting real code and debugging as I go along, and bring in the Cuda backend into the language along with the other bits and pieces. Probably I'll switch to regular syntax for the pattern matcher for the sake of uniformity, but I will leave the extra operators.

I will make my hard work be rewarded. I came this far, I might as well go all the way to the end.

I will never give up even if I die.