This should do it to start things off.
But before I do anything else, how about I do the thing that I've long thought about ever since I started work on the pattern matching compiler and get rid of CPS. It makes the compiler 2x more complicated and who knows how many more times slower for essentially no reason. I originally intended it to be for more complicated pattern matching control flow back then when pattern matching was a compile time thing only, but now it is different.

When I first implemented CPS it seemed like it would be a huge change, but it only took me like a few hours to convert a thousand lines of code.

Going back should be easier.

Right now it is 4:20pm so it is kind of late, but it might enough for a rough conversion of the pevaller. Let me do it.