Holy shit, I need a break.
I had to add a special case for turning variables into union recursive types and the effort required to figure out how to not duplicate too much code took it out of me. I am still not happy with the result.

I still feel it came out to hacky. I keep thinking there is a better way of doing this. That there is a better way of thinking about this. I cannot find it and it is so frustrating.

For the past two hours or so, honestly, I've been completely dry. I do not even feel like testing this right now.

Let me watch the Scala parser combinator video that was posted on the /r/programming sub.

https://www.youtube.com/watch?v=mARO-qchsKM

sfscala.org: Li Haoyi, FastParse: Fast, Programmable, Modern Parser-Combinators in Scala

I'd be happy to find out if these fast parsers use LMS, but probably not. When I did a little probe on /r/Scala, the only thing I got was a guy asking me what staging is.

Maybe I'll test out the changes later in the day, after I am done with lunch. I won't push myself.