"2:50pm. There is something wrong here. It is not instantiating the variable correctly. inl int_expr x = expr 0 Ah, I am such a goof. Notice the x there. Fail   "Variable "interpreter_static" not bound. Error trace on line: 15, column: 22 in file "test12".     | (.Mult a b) -> interpreter_static a * interpreter_static b Another parser error. inl rec interpreter_static x =     inl f = interpreter_static     match x with     | (.V x) -> x     | (.Add a b) -> f a + f b     | (.Mult a b) -> f a * f b Like this it works, but not without the x argument. let test13 =     "test13",     """ inl rec p = function     | (.Some x) -> p x     | .None -> 0 p (.Some, .None)     """  I simplified it a bit. How do I deal with this. I am going to have to print this out. 3:15pm. Stop browsing Reddit. Focus on solving this. pattern=[] body=Function   (("", " main_arg",     Op       (Case,        [V (" main_arg",null);         Op           (IfStatic,            [Op (VVIs,[V (" main_arg",null)],null);             Op … Why the hell is the name missing here? 3:20pm. I won't paste the thing here, but the name really is missing. 3:25pm. I had enough. Let me take a bath. ...Actually, just a bit more.         let x = inlr' name pattern body p         printfn "x=%A" x … > x=Function   (("", " main_arg",     Op       (Case,        [V (" main_arg",null);         Op           (IfStatic, Why is the name missing? > In pattern_compile_single. name= pattern=PatClauses   (Some ("test13", 2L, 13L), Why is there not a name here? 3:35pm. let rec inlr' name (args: Pattern list) body pos =     printfn "I am in inlr'. name=%s" name     let f arg body = pattern_compile_single name (with_clause pos arg body)     match args with     | x :: xs -> f x (inlr' "" xs body pos)     | [] -> body AHHHHHHhHhhhhhhaahsdwreasfrdxvdf gsxdegcxbd. I get it. Because inlr is being called without an argument it just returns body. FFFFFFFFFFFFFFFFFFFFFFFFFF..."
I sure love debugging.

This is ridiculous. Positions everwhere and names everywhere.

I need a better system for this.

I thought I was being smart by adding the Pos field into every Expr, but it is making everything in the parser segment 10x uglier. The same goes for this fucking issue with recursive function names.

Also I have not mentioned it, but compiling the pattern matcher and then doing the free variable calculation explodes the complexity of the algorithm.

It would be better to defer doing it until after the free variable calculations.

Also because Pos is a part of Expr, it is ugly as sin. It makes reading the AST 10x harder.

If I made it an Op instead, I would have the option of temporarily disabling it in the parser.

...Making a system for deferring pattern matching compilation would finally give me a reason to get rid of those refs directly in the AST that I am currently using do free variables.

At any rate, I had enough. Before I redesign the language again, let me take a break to think about it.

Getting all of this to work will no doubt be a long tow.

I tried as much as I could to avoid doing my own language, but now that I am doing it, you can be sure that I will finish it.

Everything will fall in its proper place.