I managed to overcome my mental block regarding recursion and got the fib example to stop overflowing.
Now it all makes sense. I was right that what I had previously was wrong and that the way I was approaching it is confusing.

Well, the solution was simple - it turns out I had to apply the fix to let's body instead of the else. This seems blindingly obvious now, but for the past few hours I felt like I was under a spell.

It won't happen again.

Now...the damn parser. I need to make the type keyword have the proper precedence. I'll deal with it tomorrow.

I did well today.