I am trying to get a handle on a stack overflow bug. This is one of the unwelcome side effects getting rid of CPS. I need to make sure whether there is something wrong with the semantics or whether having that destructure on every eval is what is messing with it.
How troublesome. I am going to have to find a way to look at the stack.

One of the problem that I have is that the F# debugger is so slow and unresponsive as to be useless.

For the time being, I'll try compiling the program and seeing if anything changes.