I refactored the `expr_typecheck` a little and I also modified the parser so the type keyword acts as a function. I just put it in applications as a special instead of higher on the precedence chain in expressions.
With this bit of trouble out of the way, let me recap all of the tests:

1) Does it run?
2) Does it run methods?
3) Does this method case work?
4) Does the and pattern work correctly?
5) Does basic pattern matching on static data work?
Fib) Does recursion work on the fibonacci example?
6) Does returning type level methods from methods work?
7) Do active patterns work?
8) Does the basic union type work?
9) Does the partial evaluator optimize unused match cases?
10) The worst case for partially evaluated pattern matchers.
11) Do the nested patterns work?
12) Does recursive pattern matching work on static data?
13) A more complex interpreter example on static data.
14) Does recursive pattern matching work on partially static data?

This last one in particular is a huge achievement. After struggling the whole day yesterday I finally got it to work. Before that, it took me a whole month just to figure out the design so this particular feature fits in naturally.

Each and every single one of the above tests revealed bugs in the implementation and I had to spend hours each debugging it. It would not be wrong to say that I pretty much spent the last ten days just debugging.

I think I am finally at the point where I can write real code with Spiral. I will finally find out if there is anything to this language of mine.

Now, what comes next?

I have various things in mind, before I turn to work on the Cuda backend I really should just play around with the new language...