"There is a whole bunch of stuff I need to do.
1) Dynamise
2) Arrays
3) Cuda
4) Closures
5) Type constructors for primitives. It is stupid to keep passing 0 when I mean int64 even if it works.
6) Standard syntax for pattern matching
7) Standard library functions for arrays since F#'s are so slow. I'll need to do it separately for Cuda arrays anyway.
8) Testing .NET stuff
9) Modules"

Done with Dynamize. As it turns out, implemting it revealed a bug in the implementation of recursive function names. I thought I got all the bugs out, but I was wrong.

So what is next? Adding `dyn` was easy.

Let me get the easy stuff out of the way first. I'll do #5.