https://github.com/Microsoft/visualfsharp/issues/3358
Here is the issue. To be honest, I am surprised the compiler will even attempt to generalize a statement with something inbetween. Did that get installed in some update or something. But it does work incorrectly in the REPL, so it can't be a part of some package.

Also for once, I finally found a real issue with the compiler. I expect this one to be given high priority. Now where was I? This definitely distracted me.