I finally got basic .NET interoperability with Spiral to work as test15 demonstrates. Today once again, I pretty much spent the entire day debuging and redesigning the internals. I got the partial application of method names to .NET type to work as well.
The language is quite nice now, but before I can call .NET interop complete there are still a bunch of things I need to do. I'll also need to do some light research on how to avoid having to write out the number of generic parameters at the end of the name.

I'll go into it in a later commit. I want to get something to eat right now.