For the past 12h I've been really distracted, so I feel a bit dazed right now.
https://www.cs.virginia.edu/~robins/YouAndYourResearch.pdf

Instead of programming, I've read the above talk by Richard Hamming on doing great research. It is interesting. Reading things of the sort is a waste of time, but I am dry right now so by wasting time in a directed manner I try to gather my bearings.

Since there is nothing to program right now, let me continue from the previous commit.

What my next mission will be is to figure out how to import the "System.Collections.Generic.Dictionary`2" type without `2 at the end. Probably I will just ask on SO about that. There is probably some way to do it without me having to make my own method for that.

There are a bunch of more things I need to do today.

1) I need to strip the type constructors from generic parameter application in apply.
2) I need to put in dotnet_typeof. To be honest I am not sure how this one should work.
3) I need to put in arrays.

For lack of direction, let me just work on the stuff in the order I've written them out.