I've thought about it for a while and let me move on from #8. I won't put it all the .NET interop. Also based on the comments to the SO question I'll have the full name with the number of parameters written out by design.
Here is what will be missing:

1) bytef
2) typeof

So I won't be able to create Arrays using Create instance and I won't be able to use the TryGetValue from the dictionary class. Instead, I will have the arrays built in and I will use the get_item property instead.

I also won't bother checking whether a method is public or private. .NET interop by itself is not hugely important and I am not interested in diving deeper into the innards of .NET reflection.

If I start putting in everything the .NET supports I will never be done with the language.

...No actually, with respect to methods, it might be possible to check a flag rather than setting the bindingFlags...let me try it.

...It works. I guess checking whether the thing is public is easy. Now I have that extra bit of polish.

Good.

Of course, unlike in F#, I won't allow the creation of .NET classes or anything like that. That would be way beyond what I am willing work on in the language.

I already have a better object system anyway.

...So that is all good.

With this I am done with .NET interop for the time being and maybe, but unlikely for good.

2) Arrays
3) Cuda
4) Closures
6) Standard syntax for pattern matching
7) Standard library functions for arrays since F#'s are so slow. I'll need to do them separately for Cuda arrays anyway.
9) Modules

Here is the list I've been working through.

As per plan, now that I have access to the console class, how about I do some very simple HackerRank exercises just to get my feet wet in the new language.

Then I will slowly put in arrays and get to work on #7.

So far I've been working on this language for 5 months and 3 weeks. That is an enormous emount of time to be working on a project and finally I can do something useful with it.

I think it is time I finally play around with it a little. I deserve this much.