Today I've been doing some more imagination programming. Being able to only propagate variables forward is a massive pain in the ass, but I've been holding up so far.
```
inl tuple2 a b s =
    match a s with
    | .Succ, a ->
        match b s with
        | .Succ, b -> .Succ, (a, b)
        | x -> x
    | x -> x
    : ...
```

This fragment is an exception. Without unification, it might turn out that I can't type it.

The reason for that is that I can only get the full type in the true branch, but not in the other ones.

...

Agh. No, I can't think of anything.

What I am going to have to do is write the parser combinator library in continuation passing style then.

I've spent the whole day writing the parsers library, but the lesson is that type inference is really a helpful thing. Maybe HM + typeclasses would be something I need in the end, but let me try this for now.