inl tuple2_cps a b s ret =     inl get_type f =         type (f s <| function                 | .Succ, a -> a                 | _ -> bottom)     inl a_ty = get_type a     inl b_ty = get_type b     inl retf = List (a_ty, b_ty)
    a s <| function
        | .Succ, a ->
            b s <| function
                | .Succ, b -> retf (.Succ, (a, b))
                | x -> retf x
        | x -> retf x

met rec many_cps p s ret =
    inl state = s.pos
    inl x_ty =
        type (p s <| function
                | .Succ, x -> x
                | _ -> bottom)
    inl retf = type (List x) >> ret
    p s <| function
        | .Succ, (^dyn x) when state < s.pos ->
            many_cps p s <| function
                | .Succ, xs -> retf (.Succ, (.ListCons, (x, xs)))
                | x -> retf x
        | .Succ, _ when state = s.pos -> retf (.FatalFail, "Many parser succeeded without changing the parser state. Unless the computation had been aborted, the parser would have gone into an infinite loop.")
        | .Fail, _ when state = s.pos -> retf (.Succ, .ListNil)
        | .Fail, _ -> retf (.Fail, (pos, "many"))
        | .FatalFail, _ -> retf x

This should be it, but…

What do I do about the return type annotation in many_cps? It won't be List x, but List x applied to ret.

Agghh, do I make a fake instance of a List or something?

This is pretty discouraging. But in the worst case, I can always make another language and then compile it to this one. But how would I even monomorphise the polymorphic types?