This kind of code stabilized my mental state. Yeah, this is the true way.
I have to say, after this experience, I see type systems in a whole new light. What I am doing above is simply emulating what a type system would do – keep track of generic variables and apply them as needed.

I could not reach it yesterday, because I was hung up on making a break with the old way of doing things, but as it turns out, the old way is an integral part of reality. Notice that in tuple2 I am not applying the types to the two parser arguments. That corresponds to the emulation of first rank polymorphism.

If in read_int, the pint and spaces parser had generic arguments, I would have applied them there. That is the way to do it.

I've thought a lot yesterday and have essentially concluded that with a type system, I could automate the keeping track and inference of generic variables, but as I made those conclusions, I understood how to do it by hand. And moreso than that – it is not like what I did just now was ever beyond my ability, what I realized is that this is the correct way to do it.

Plunging into the innards of types can't possibly be the way to do it, instead propagating the types as the type systems do it is definitely right.

...Now, I'll still play around with LMS and today my mission will be to pore over the posts by that Indian Scala guy.